requirements.txt
streamlit
tensorflow>=2.18.0
opencv-python
pillow
numpy
pandas
pymongo
bcrypt
python-dotenv
scikit-learn
matplotlib
plotly
fpdf
